fix some dataset may not have peptides

+++ R/txt2binding.R
@@ -36,23 +36,27 @@ txt2pep <- function(annovar_path,txt_path,
 
   variants <- names(dt)[seq(2,length(dt),by=2)]
   variants <- variants[grepl("protein-altering",variants)]
-  pos_alter <- stringr::str_extract_all(variants,"p[.].+ protein-altering") %>% gsub(" protein-altering","",.)
-  cdna <- stringr::str_extract_all(variants,"c[.].+ p[.]") %>% gsub(" p[.]","",.)
-  lines <- stringr::str_extract_all(variants,"line.+ NM_") %>% gsub(" NM_","",.) %>%
-    gsub("line","",.) %>% as.numeric()
-  ENST <- stringr::str_extract_all(variants,"line.+ NM_[0-9]+") %>% gsub("line.+ ","",.)##extract transcripts
+  if (length(variants) == 0){
+    return(NULL)
+  }else{
+    pos_alter <- stringr::str_extract_all(variants,"p[.].+ protein-altering") %>% gsub(" protein-altering","",.)
+    cdna <- stringr::str_extract_all(variants,"c[.].+ p[.]") %>% gsub(" p[.]","",.)
+    lines <- stringr::str_extract_all(variants,"line.+ NM_") %>% gsub(" NM_","",.) %>%
+      gsub("line","",.) %>% as.numeric()
+    ENST <- stringr::str_extract_all(variants,"line.+ NM_[0-9]+") %>% gsub("line.+ ","",.)##extract transcripts
 
-  mut_need <- mut[lines,]
-  pep_seq_mt <- dt[sapply(variants,function(x){which(names(dt)==x)}) %>% unname()] %>% as.character()
-  pep_seq_wt <- dt[(sapply(variants,function(x){which(names(dt)==x)}) %>% unname())-1] %>% as.character()
-  mut_need$seq_wt <- pep_seq_wt
-  mut_need$seq_mt <- pep_seq_mt
-  mut_need$pos_alter <- pos_alter
-  mut_need$cdna <- cdna
-  mut_need$transcript <- ENST
-  files_exist <- c(list.files(temp_dir,pattern = "myanno",full.names = T))##remove all temp files
-  file.remove(files_exist)
-  return(mut_need)
+    mut_need <- mut[lines,]
+    pep_seq_mt <- dt[sapply(variants,function(x){which(names(dt)==x)}) %>% unname()] %>% as.character()
+    pep_seq_wt <- dt[(sapply(variants,function(x){which(names(dt)==x)}) %>% unname())-1] %>% as.character()
+    mut_need$seq_wt <- pep_seq_wt
+    mut_need$seq_mt <- pep_seq_mt
+    mut_need$pos_alter <- pos_alter
+    mut_need$cdna <- cdna
+    mut_need$transcript <- ENST
+    files_exist <- c(list.files(temp_dir,pattern = "myanno",full.names = T))##remove all temp files
+    file.remove(files_exist)
+    return(mut_need)
+  }
 }
 
 #' Convert TXT file to mutated peptides and extract proper range of amino acids
@@ -76,24 +80,28 @@ txt2seq <- function(annovar_path,txt_path,
   pep <- txt2pep(annovar_path = annovar_path,txt_path = txt_path,
                  genome_version = genome_version,tmp_dir=tmp_dir,num_thread=num_thread)
 
-  pep <- pep %>%
-    mutate(pos = stringr::str_extract(pos_alter,"[0-9]+") %>% as.numeric()) %>%
-    rowwise() %>%
-    mutate(indel=ifelse(ref != "-" & alt != "-" & nchar(ref) == 1 & nchar(alt) == 1,"FALSE","TRUE")) %>%
-    as.data.frame()
-  ext_seqs_mt <- mapply(MHCbinding::extractSeq,seq=pep$seq_mt,pos=as.numeric(pep$pos),
-                        len=as.numeric(len),indel=pep$indel) %>% unname()
-  ext_seqs_wt <- mapply(MHCbinding::extractSeq,seq=pep$seq_wt,pos=as.numeric(pep$pos),
-                        len=as.numeric(len),indel=pep$indel) %>% unname()
-  ##remove the last star
-  a <-  substr(ext_seqs_mt,nchar(ext_seqs_mt),nchar(ext_seqs_mt))=="*"
-  b <-  substr(ext_seqs_wt,nchar(ext_seqs_wt),nchar(ext_seqs_wt))=="*"
-  ext_seqs_mt[a] <- gsub("\\*","",ext_seqs_mt[a])
-  ext_seqs_wt[b] <- gsub("\\*","",ext_seqs_wt[b])
+  if (is.null(pep)){
+    return(NULL)
+  }else{
+    pep <- pep %>%
+      mutate(pos = stringr::str_extract(pos_alter,"[0-9]+") %>% as.numeric()) %>%
+      rowwise() %>%
+      mutate(indel=ifelse(ref != "-" & alt != "-" & nchar(ref) == 1 & nchar(alt) == 1,"FALSE","TRUE")) %>%
+      as.data.frame()
+    ext_seqs_mt <- mapply(MHCbinding::extractSeq,seq=pep$seq_mt,pos=as.numeric(pep$pos),
+                          len=as.numeric(len),indel=pep$indel) %>% unname()
+    ext_seqs_wt <- mapply(MHCbinding::extractSeq,seq=pep$seq_wt,pos=as.numeric(pep$pos),
+                          len=as.numeric(len),indel=pep$indel) %>% unname()
+    ##remove the last star
+    a <-  substr(ext_seqs_mt,nchar(ext_seqs_mt),nchar(ext_seqs_mt))=="*"
+    b <-  substr(ext_seqs_wt,nchar(ext_seqs_wt),nchar(ext_seqs_wt))=="*"
+    ext_seqs_mt[a] <- gsub("\\*","",ext_seqs_mt[a])
+    ext_seqs_wt[b] <- gsub("\\*","",ext_seqs_wt[b])
 
-  pep$ext_seqs_wt <- ext_seqs_wt
-  pep$ext_seqs_mt <- ext_seqs_mt
-  return(pep)
+    pep$ext_seqs_wt <- ext_seqs_wt
+    pep$ext_seqs_mt <- ext_seqs_mt
+    return(pep)
+  }
 }
 
 #' Predict peptide MHC binding based on TXT file.
@@ -132,52 +140,57 @@ txt2binding <- function(get_method=c("api","client"),annovar_path,txt_path,
                         len = pep_length[i],tmp_dir=tmp_dir,num_thread=num_thread)
   }
   res <- dplyr::bind_rows(res,.id = "predicted_length")
-  res <- res[which(nchar(res$ext_seqs_mt) >= as.numeric(res$predicted_length)),]
-  res <- res %>%
-    select(-seq_mt,-seq_wt) %>%
-    group_by(predicted_length,chr,start,end,ref,alt,indel,ext_seqs_mt,ext_seqs_wt) %>%
-    summarise(cdna=paste(cdna,collapse = ","),
-              transcript=paste(transcript,collapse = ","),
-              pos_alter=paste(pos_alter,collapse = ","))
+  if(nrow(res)==0){
+    message("There is no peptides!")
+    return(NULL)
+  }else{
+    res <- res[which(nchar(res$ext_seqs_mt) >= as.numeric(res$predicted_length)),]
+    res <- res %>%
+      select(-seq_mt,-seq_wt) %>%
+      group_by(predicted_length,chr,start,end,ref,alt,indel,ext_seqs_mt,ext_seqs_wt) %>%
+      summarise(cdna=paste(cdna,collapse = ","),
+                transcript=paste(transcript,collapse = ","),
+                pos_alter=paste(pos_alter,collapse = ","))
 
-  pep_length <- unique(res$predicted_length)
-  pre_res_mt <- vector("list",length = length(pep_length))
-  pre_res_wt <- vector("list",length = length(pep_length))
-  names(pre_res_mt) <- pep_length
-  names(pre_res_wt) <- pep_length
-  for (i in seq_along(pre_res_mt)){
-    pep_mt <- res[res$predicted_length == names(pre_res_mt)[i],"ext_seqs_mt"]
-    pre_res_mt[[i]] <- MHCbinding:::general_mhcbinding(get_method = get_method,mhc_type = mhc_type, length = pep_length[i],
-                                                    allele = allele,pre_method = pre_method, peptide = pep_mt$ext_seqs_mt,client_path = client_path,
-                                                    tmp_dir=tmp_dir)
-    pep_wt <- res[res$predicted_length == names(pre_res_wt)[i],"ext_seqs_wt"]
-    pre_res_wt[[i]] <- MHCbinding:::general_mhcbinding(get_method = get_method,mhc_type = mhc_type, length = pep_length[i],
-                                                       allele = allele,pre_method = pre_method, peptide = pep_wt$ext_seqs_wt,client_path = client_path,
-                                                       tmp_dir=tmp_dir)
+    pep_length <- unique(res$predicted_length)
+    pre_res_mt <- vector("list",length = length(pep_length))
+    pre_res_wt <- vector("list",length = length(pep_length))
+    names(pre_res_mt) <- pep_length
+    names(pre_res_wt) <- pep_length
+    for (i in seq_along(pre_res_mt)){
+      pep_mt <- res[res$predicted_length == names(pre_res_mt)[i],"ext_seqs_mt"]
+      pre_res_mt[[i]] <- MHCbinding:::general_mhcbinding(get_method = get_method,mhc_type = mhc_type, length = pep_length[i],
+                                                         allele = allele,pre_method = pre_method, peptide = pep_mt$ext_seqs_mt,client_path = client_path,
+                                                         tmp_dir=tmp_dir)
+      pep_wt <- res[res$predicted_length == names(pre_res_wt)[i],"ext_seqs_wt"]
+      pre_res_wt[[i]] <- MHCbinding:::general_mhcbinding(get_method = get_method,mhc_type = mhc_type, length = pep_length[i],
+                                                         allele = allele,pre_method = pre_method, peptide = pep_wt$ext_seqs_wt,client_path = client_path,
+                                                         tmp_dir=tmp_dir)
+    }
+
+    pre_res_mt <- dplyr::bind_rows(pre_res_mt)
+    pre_res_wt <- dplyr::bind_rows(pre_res_wt)
+
+    res <- res %>%
+      group_by(predicted_length) %>%
+      mutate(seq_num=row_number()) %>%
+      mutate(index=paste(predicted_length,seq_num,sep = ":"))
+    pre_res_mt <- pre_res_mt %>% mutate(index=paste(length,seq_num,sep = ":"))
+    pre_res_mt <- left_join(
+      pre_res_mt %>% rename(pep_start=start,pep_end=end),
+      res %>% ungroup() %>% select(chr,start,end,ref,alt,index,ext_seqs_mt,ext_seqs_wt,pos_alter,cdna,transcript)
+    ) %>% select(-index,-seq_num) %>% select(chr,start,end,ref,alt,ext_seqs_mt,ext_seqs_wt,
+                                             pos_alter,cdna,transcript,everything())
+    pre_res_mt$peptide_wt <- substr(pre_res_mt$ext_seqs_wt,pre_res_mt$pep_start,pre_res_mt$pep_end)
+
+    pre_res <- left_join(
+      pre_res_mt %>% mutate(index=paste(allele,peptide_wt,sep = ":")),
+      pre_res_wt %>% select(allele,peptide,ic50,rank) %>%
+        distinct_all(.keep_all = T) %>%
+        mutate(index=paste(allele,peptide,sep = ":")) %>%
+        select(index,ic50,rank) %>% rename(wt_ic50=ic50,wt_rank=rank)
+    ) %>% select(-index)
+    return(pre_res)
   }
-
-  pre_res_mt <- dplyr::bind_rows(pre_res_mt)
-  pre_res_wt <- dplyr::bind_rows(pre_res_wt)
-
-  res <- res %>%
-    group_by(predicted_length) %>%
-    mutate(seq_num=row_number()) %>%
-    mutate(index=paste(predicted_length,seq_num,sep = ":"))
-  pre_res_mt <- pre_res_mt %>% mutate(index=paste(length,seq_num,sep = ":"))
-  pre_res_mt <- left_join(
-    pre_res_mt %>% rename(pep_start=start,pep_end=end),
-    res %>% ungroup() %>% select(chr,start,end,ref,alt,index,ext_seqs_mt,ext_seqs_wt,pos_alter,cdna,transcript)
-  ) %>% select(-index,-seq_num) %>% select(chr,start,end,ref,alt,ext_seqs_mt,ext_seqs_wt,
-                                           pos_alter,cdna,transcript,everything())
-  pre_res_mt$peptide_wt <- substr(pre_res_mt$ext_seqs_wt,pre_res_mt$pep_start,pre_res_mt$pep_end)
-
-  pre_res <- left_join(
-    pre_res_mt %>% mutate(index=paste(allele,peptide_wt,sep = ":")),
-    pre_res_wt %>% select(allele,peptide,ic50,rank) %>%
-      distinct_all(.keep_all = T) %>%
-      mutate(index=paste(allele,peptide,sep = ":")) %>%
-      select(index,ic50,rank) %>% rename(wt_ic50=ic50,wt_rank=rank)
-  ) %>% select(-index)
-  return(pre_res)
 }
